+++ backend/src/session_photo/session_photo.service.ts
@@ -3,10 +3,23 @@ import { GetSessionPhotoDto } from './dto/get-session_photo.dto';
 import { PrismaService } from '../prisma/prisma.service';
 import { CreateSessionPhotoDto } from './dto/create-session_photo.dto';
 import { UpdateSessionPhotoDto } from './dto/update-session.dto';
+import { CurrentUserProvider } from 'src/current-user/current-user.provider';
+import { ForbiddenException } from 'src/exception/service-exception';
 
 @Injectable()
 export class SessionPhotoService {
-  constructor(private prisma: PrismaService) {}
+  constructor(
+    private prisma: PrismaService,
+    private currentUserProvider: CurrentUserProvider,
+  ) {}
+
+  private get userId(): string {
+    return this.currentUserProvider.user.userId;
+  }
+
+  private async authorize(id: string): Promise<void> {
+    this.findOne(id);
+  }
 
   async create(
     sessionId: string,
@@ -26,18 +39,35 @@ export class SessionPhotoService {
     });
   }
 
-  async findOne(id: string): Promise<GetSessionPhotoDto | null> {
-    return await this.prisma.sessionPhoto.findUnique({
+  async findOne(id: string): Promise<GetSessionPhotoDto> {
+    const sessionPhotoWithProject = await this.prisma.sessionPhoto.findUnique({
       where: {
         id: id,
       },
+      include: {
+        session: {
+          include: {
+            project: true,
+          },
+        },
+      },
     });
+
+    if (sessionPhotoWithProject.session.project.userId != this.userId) {
+      throw new ForbiddenException();
+    }
+
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    const { session, ...sessionPhoto } = sessionPhotoWithProject;
+    return sessionPhoto;
   }
 
   async update(
     id: string,
     updateSessionPhotoDto: UpdateSessionPhotoDto,
   ): Promise<void> {
+    await this.authorize(id);
+
     await this.prisma.user.update({
       where: {
         id: id,
@@ -49,6 +79,8 @@ export class SessionPhotoService {
   }
 
   async remove(id: string): Promise<void> {
+    await this.authorize(id);
+
     await this.prisma.sessionPhoto.delete({
       where: { id: id },
     });
